Nueva funcionalidad para generar listado de ofertas con sublistados por tipo de productos con cantidad y subtotal por c/u

+++ AsistenteCompras_API/AsistenteCompras_Entities/Entities/Producto.cs
@@ -16,7 +16,7 @@ public partial class Producto
 
     public string Imagen { get; set; } = null!;
 
-    public double Peso { get; set; }
+    public int Peso { get; set; }
 
     public int Unidades { get; set; }
